+++ crates/aero-machine/tests/boot_int10_aerogpu_vbe_115_sets_mode.rs
@@ -69,7 +69,7 @@ fn boot_int10_aerogpu_vbe_115_sets_mode_and_lfb_is_visible_via_bar1() {
 
     // The BIOS should report the LFB base as BAR1_BASE + 0x40000 (after the 256KiB legacy planar
     // VGA reservation).
-    let lfb_base = u64::from(m.vbe_lfb_base());
+    let lfb_base = m.vbe_lfb_base();
     assert_eq!(lfb_base, bar1_base + VBE_LFB_OFFSET as u64);
 
     // Write a red pixel at (0,0) in VBE packed-pixel B,G,R,X format.
